Extract DKG roster/auth layer into keep-frost-net::dkg, shared by CLI
